==> 03. Data filtering.ipynb <==
# Data filtering

In dit Notebook wordt getoond hoe je een selectie kan maken maken voor verder onderzoek. 

Gezien de beperkte grootte van Binder qua opslagruimte, richten we ons op een vergelijking tussen artikelen over fietsen in regionale en landelijke dagbladen. 

De eerst stap is het importeren van de pandas package:

import pandas as pd

Daarna laden we het dataframe

data = pd.read_csv('data/data_new.csv', index_col = 0)

En we kijken of alles goed is gegaan:

data.head(2)

Als eerste maken we een nieuw dataframe aan, genaamd `fiets_df`.

In de eerste stap filteren we op basis van de kolom `category` alleen de rijen `fiets` eruit. 

fiets_df = data[data['category'] == 'fiets']

Tijdens de data exploratie zagen we dat de transcriptie voor de familieberichten en de advertenties gemiddeld genomen erg slecht was. We willen daarom alleen maar de artikelen overhouden. 

Ook hiervoor maken we een filter.

**Let op!** Dit nieuwe filteren passen we toe op het zojuist gemaakte dataframe `fiets_df`

fiets_df = fiets_df[fiets_df['type'] == 'artikel']

len(fiets_df)

Aangezien we in het volgende dataframe tekstuele analyses willen doen, willen we niet dat de kwaliteit van de tekst te slecht is. Daarom filteren we alle artikelen eruit met een DL score lager dan 70%. 

fiets_df = fiets_df[fiets_df['DL score'] >=70]

En we slaan deze set op voor verder gebruik

fiets_df.to_csv('data/fiets.csv')
