==> 06. Part-of-speech example.ipynb <==
# Extra:  Part of speech tagging
Part of Speech (POS) tagging is een NLP activiteit waarbij elk woord in een tekst een grammaticale categorie krijgt toegewezen (voornaamwoord, werkwoord, naamwoord, etc).
Het toevoegen van de POS maakt het makkelijker om (automatisch) de structuur en betekenis van de tekst te doorgronden.
Door POS te gebruiken is het relatief makkeijk het onderwerp van de tekst te vinden (bijvoorbeeld door de zelfstandige naamwoorden of de grammaticale onderwerpen te bekijken. 

Hier een voorbeeld van een paar POS categorieën die Spacy herkent. 

| POS    | Description |Examples |
| -------- | ------- |------- |
| ADV  | adverb | very, tomorrow, down, where, there  |
| NOUN | noun  | girl, cat, tree, air, beauty |
| VERB |  verb | run, runs, running, eat, ate, eating |
| PRON | pronoun  | I, you, he, she, myself, themselves, somebody |
| SCONJ | subordinating conjunction  | if, while, that |
| NUM | numeral  | 1, 2017, one, seventy-seven, IV, MMXIV |
| INTJ | interjection | psst, ouch, bravo, hello |
| X | other  | sfpksdpsxmsa |




In onderstaande voorbeeld gebruiken we POS om te kijken of er een verschil is in de opbouw van teksten tussen regionale/lokale en landelijke kranten. We doen dit door te kijken naar het gemiddeld aantal voorkomens per woordcategorie per tekst. 

Als eerste stap laden we de benodigde packages. 

import pandas as pd
import spacy
from spacy import displacy
import pickle

En we openen ons opgeslagen model, met daarin de kolom `doc` die de POS informatie bevat. 

fiets_nlp = pd.read_pickle("data/fiets_nlp.pkl")

fiets_nlp.head(2)

Om de POS-tags uit de `doc` kolom te halen, is onderstaande functie geschreven:

def get_pos(doc, pos_tag):
    return [token for token in doc if token.pos_ == pos_tag]

Hierna doorloop je voor elke POS tag de volgende drie stappen:

Stap 1: maak een nieuwe kolom aan met de woordcategorie die je wilt extraheren. 

fiets_nlp['noun'] = fiets_nlp['doc'].apply(lambda x: get_pos(x, 'NOUN'))

Stap 2: Maak een nieuwe kolom aan met het aantal woorden van deze woordcategorie. Dit doe je door het aantal woorden in de zojuist gemaakte kolom te tellen. 

fiets_nlp['noun_count'] = fiets_nlp['noun'].apply(lambda x: len(x))

Stap 3: Bereken in een nieuwe kolom het percentage woorden van deze categorie per artikel

fiets_nlp['noun_perc'] = (fiets_nlp['noun_count'] / fiets_nlp['length']) * 100

In het dataframe zie je nu drie nieuwe kolommen. 

fiets_nlp.head(2)

Dezelfde stappen kun je ook voor andere woordcategoriën doen. Hieronder zie je twee voorbeelden:

fiets_nlp['adv'] = fiets_nlp['doc'].apply(lambda x: get_pos(x, 'ADV'))
fiets_nlp['adv_count'] = fiets_nlp['adv'].apply(lambda x: len(x))
fiets_nlp['adv_perc'] = (fiets_nlp['adv_count'] / fiets_nlp['length']) * 100

fiets_nlp['pron'] = fiets_nlp['doc'].apply(lambda x: get_pos(x, 'PRON'))
fiets_nlp['pron_count'] = fiets_nlp['pron'].apply(lambda x: len(x))
fiets_nlp['pron_perc'] = (fiets_nlp['pron_count'] / fiets_nlp['length']) * 100

Nu je van meerdere woordsoorten het percentage per artikel hebt, kun je bekijken of er een verschil is in woordgebruik tussen regionale en landelijke kranten. 

We maken als eerst een variabele aan met daarin alle kolomnamen waarin we het percentage hebben berekend:

columns = ['noun_perc', 'adv_perc', 'pron_perc']

En vervolgens kun je met behulp van een `groupby` functie op deze kolommen kijken of er een verschil is in voorkomen tussen de regionale/lokale en de landelijke kranten

fiets_nlp.groupby(['spatial'])[columns].mean()

In plaats van `mean()` kan je ook `median()` gebruiken om de mediaan te bekijken. 

fiets_nlp.groupby(['spatial'])[columns].median()

Je kunt zelf andere woordsoorten toevoegen om de vergelijking uit te breiden.
